+++ src/Exception/EnumNotRegisteredException.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 /*
  * This file is part of the Enumer package.
  *
- * (c) MarfaTech <https://wakeapp.ru>
+ * (c) MarfaTech <https://marfa-tech.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -13,6 +13,6 @@ declare(strict_types=1);
 
 namespace MarfaTech\Component\Enumer\Exception;
 
-class EnumNotRegisteredException extends MarfatechEnumerException
+class EnumNotRegisteredException extends MarfaTechEnumerException
 {
 }